split helper functions part 10
