+++ InitialWorkflowModule.lua
@@ -74,12 +74,6 @@ if not Helper.CheckApiVersion() then
   return
 end
 
-local Env =
-{
-  InstallModuleEventRegistered = false,
-  InstallModuleDone = false,
-}
-
 LogHelper.Info(string.format(_("script executed from path %s"), ScriptFilePath()))
 LogHelper.Info(string.format(_("script translation files in %s"), localePath))
 LogHelper.Info(_("script outputs are in English"))
@@ -113,292 +107,24 @@ WorkflowSteps.Init(dt, LogHelper, Helper, EventHelper, GuiTranslation, Workflow,
 
 -- init ./Modules/Buttons.lua
 local WorkflowButtons = require 'Modules.WorkflowButtons'
-WorkflowButtons.Init(dt, LogHelper, Helper, EventHelper, GuiTranslation, Workflow, GuiAction, WidgetStack, ScriptFilePath())
+WorkflowButtons.Init(dt, LogHelper, Helper, EventHelper, GuiTranslation, Workflow, GuiAction, WidgetStack,
+ScriptFilePath())
 
+-- init widget controls
 WorkflowSteps.CreateWorkflowSteps()
 WorkflowButtons.CreateWorkflowButtons()
 
----------------------------------------------------------------
---[[
-
-      IMPLEMENTATION OF WIDGET FRAME
-
-      Create main widget. Collect buttons and comboboxes.
-
-    ]]
----------------------------------------------------------------
-
-local ResetAllCommonMainSettingsWidget
-ResetAllCommonMainSettingsWidget = dt.new_widget('combobox')
-    {
-      changed_callback = function()
-        local selection = ResetAllCommonMainSettingsWidget.value
-
-        if (selection ~= _dt("all common settings")) then
-          for i, step in ipairs(Workflow.ModuleSteps) do
-            if step.WidgetStackValue == WidgetStack.Settings then
-              if (step ~= StepTimeout) then
-                if (selection == _("default")) then
-                  LogHelper.Info(step.Label)
-                  step:EnableDefaultStepConfiguation()
-                end
-              end
-            end
-          end
-
-          -- reset to default selection
-          ResetAllCommonMainSettingsWidget.value = 1
-        end
-      end,
-      label = ' ',
-      tooltip = Helper.Wordwrap(_(
-        "Configure all following common settings: Reset all steps to default configurations. These settings are applied during script run, if corresponding step is enabled.")),
-      table.unpack({ _dt("all common settings"), _("default") })
-    }
-
-local ResetAllModuleBasicSettingsWidget
-ResetAllModuleBasicSettingsWidget = dt.new_widget('combobox')
-    {
-      changed_callback = function()
-        local selection = ResetAllModuleBasicSettingsWidget.value
-
-        if (selection ~= _dt("all module basics")) then
-          for i, step in ipairs(Workflow.ModuleSteps) do
-            if step.WidgetStackValue == WidgetStack.Modules then
-              if (step ~= StepTimeout) then
-                if (selection == _("default")) then
-                  step:EnableDefaultBasicConfiguation()
-                else
-                  step:SetWidgetBasicValue(selection)
-                end
-              end
-            end
-          end
-
-          -- reset to default selection
-          ResetAllModuleBasicSettingsWidget.value = 1
-        end
-      end,
-      label = ' ',
-      tooltip = Helper.Wordwrap(_(
-        "Configure all module settings: a) Select default value. b) Ignore this step / module and do nothing at all. c) Enable corresponding module and set selected module configuration. d) Reset the module and set selected module configuration. e) Disable module and keep it unchanged.")),
-      table.unpack({ _dt("all module basics"), _("default"), _("ignore"), _("enable"), _("reset"), _("disable") })
-    }
-
-local ResetAllModuleMainSettingsWidget
-ResetAllModuleMainSettingsWidget = dt.new_widget('combobox')
-    {
-      changed_callback = function()
-        local selection = ResetAllModuleMainSettingsWidget.value
-
-        if (selection ~= _dt("all module settings")) then
-          for i, step in ipairs(Workflow.ModuleSteps) do
-            if step.WidgetStackValue == WidgetStack.Modules then
-              if (step ~= StepTimeout) then
-                if (selection == _("default")) then
-                  step:EnableDefaultStepConfiguation()
-                elseif (selection == _("unchanged")) then
-                  -- choose 'unchanged' step setting
-                  -- configuration keeps unchanged during script execution
-                  step.Widget.value = step.WidgetUnchangedStepConfigurationValue
-                end
-              end
-            end
-          end
-
-          -- reset to default selection
-          ResetAllModuleMainSettingsWidget.value = 1
-        end
-      end,
-      label = ' ',
-      tooltip = Helper.Wordwrap(_(
-        "Configure all module settings: Keep all modules unchanged or enable all default configurations. These configurations are set, if you choose 'reset' or 'enable' as basic setting.")),
-      table.unpack({ _dt("all module settings"), _("default"), _("unchanged") })
-    }
-
-----------------------------------------------------------
-
--- TEST button: Special buttons, used to perform module tests.
-local function GetWidgetTestButtons(widgets)
-  if (ButtonModuleTest) then
-    LogHelper.Info(_("insert test button widget"))
-    table.insert(widgets,
-      dt.new_widget('box')
-      {
-        orientation = 'horizontal',
-        ButtonModuleTest.Widget,
-        ButtonModuleTestCustomCode.Widget
-      }
-    )
-  end
-end
-
--- add buttons to simplify some manual steps
-local function GetWidgetOverallButtons(widgets)
-  local buttonColumn1 = {}
-  local buttonColumn2 = {}
-  local buttonColumn3 = {}
-
-  -- add overall function buttons, first column
-  table.insert(buttonColumn1, ButtonEnableRotateAndPerspective.Widget)
-  table.insert(buttonColumn1, ButtonRunSelectedSteps.Widget)
-
-  -- ... second column
-  table.insert(buttonColumn2, ButtonEnableCrop.Widget)
-  table.insert(buttonColumn2, ButtonShowWidgetStackModules.Widget)
-
-  -- ... third column
-  table.insert(buttonColumn3, ButtonMidToneExposure.Widget)
-  table.insert(buttonColumn3, ButtonShowWidgetStackSettings.Widget)
-
-  table.insert(widgets,
-    dt.new_widget('box')
-    {
-      orientation = 'horizontal',
-
-      dt.new_widget('box') {
-        orientation = 'vertical',
-        table.unpack(buttonColumn1),
-      },
-
-      dt.new_widget('box') {
-        orientation = 'vertical',
-        table.unpack(buttonColumn2),
-      },
-
-      dt.new_widget('box') {
-        orientation = 'vertical',
-        table.unpack(buttonColumn3),
-      },
-    }
-  )
-end
-
--- add comboboxes to configure workflow steps
-local function GetWidgetModulesBox()
-  -- add workflow step controls
-  -- first column: label widgets
-  -- second column: basic widgets (reset, enable, ignore...)
-  -- third column: step configuration combobox widgets
-
-  local labelWidgetsModules = {}
-  local basicWidgetsModules = {}
-  local comboBoxWidgetsModules = {}
-
-  -- add comboboxes to configure or reset all configurations
-
-  table.insert(labelWidgetsModules, dt.new_widget('label') { label = ' ' })
-  table.insert(basicWidgetsModules, ResetAllModuleBasicSettingsWidget)
-  table.insert(comboBoxWidgetsModules, ResetAllModuleMainSettingsWidget)
-
-  table.insert(labelWidgetsModules, dt.new_widget('label') { label = ' ' })
-  table.insert(basicWidgetsModules, dt.new_widget('label') { label = ' ' })
-  table.insert(comboBoxWidgetsModules, dt.new_widget('label') { label = ' ' })
-
-  for i, step in ipairs(Workflow.ModuleSteps) do
-    if step.WidgetStackValue == WidgetStack.Modules then
-      table.insert(labelWidgetsModules, step.WidgetLabel)
-      table.insert(basicWidgetsModules, step.WidgetBasic)
-      table.insert(comboBoxWidgetsModules, step.Widget)
-    end
-  end
-
-  local box = dt.new_widget('box')
-      {
-        orientation = 'horizontal',
-
-        dt.new_widget('box') {
-          orientation = 'vertical',
-          table.unpack(labelWidgetsModules),
-        },
-
-        dt.new_widget('box') {
-          orientation = 'vertical',
-          table.unpack(basicWidgetsModules),
-        },
-
-        dt.new_widget('box') {
-          orientation = 'vertical',
-          table.unpack(comboBoxWidgetsModules),
-        },
-      }
-
-  return box
-end
-
--- add comboboxes to configure workflow steps
-local function GetWidgetSettingsBox()
-  -- add setting controls
-
-  local labelWidgetsSettings = {}
-  local basicWidgetsSettings = {}
-  local comboBoxWidgetsSettings = {}
-
-  -- add comboboxes to configure or reset all configurations
-
-  table.insert(labelWidgetsSettings, dt.new_widget('label') { label = ' ' })
-  table.insert(basicWidgetsSettings, dt.new_widget('label') { label = ' ' })
-  table.insert(comboBoxWidgetsSettings, ResetAllCommonMainSettingsWidget)
-
-  table.insert(labelWidgetsSettings, dt.new_widget('label') { label = ' ' })
-  table.insert(basicWidgetsSettings, dt.new_widget('label') { label = ' ' })
-  table.insert(comboBoxWidgetsSettings, dt.new_widget('label') { label = ' ' })
-
-  -- add setting controls
-
-  for i, step in ipairs(Workflow.ModuleSteps) do
-    if step.WidgetStackValue == WidgetStack.Settings then
-      table.insert(labelWidgetsSettings, step.WidgetLabel)
-      table.insert(basicWidgetsSettings, step.WidgetBasic)
-      table.insert(comboBoxWidgetsSettings, step.Widget)
-    end
-  end
-
-  local box = dt.new_widget('box')
-      {
-        orientation = 'horizontal',
-
-        dt.new_widget('box') {
-          orientation = 'vertical',
-          table.unpack(labelWidgetsSettings),
-        },
-
-        dt.new_widget('box') {
-          orientation = 'vertical',
-          table.unpack(basicWidgetsSettings),
-        },
-
-        dt.new_widget('box') {
-          orientation = 'vertical',
-          table.unpack(comboBoxWidgetsSettings),
-        },
-      }
-
-  return box
-end
-
--- collect all widgets to be displayed within the module
--- collect buttons and comboboxes (basic and configuration)
--- the order in the GUI is the same as the order of declaration in this function.
-local function GetWidgets()
-  local widgets = {}
-
-  GetWidgetTestButtons(widgets)
-  GetWidgetOverallButtons(widgets)
-  local boxModules = GetWidgetModulesBox()
-  local boxSettings = GetWidgetSettingsBox()
-
-  -- create widget stack
-  WidgetStack.Stack[WidgetStack.Modules] = boxModules
-  WidgetStack.Stack[WidgetStack.Settings] = boxSettings
-  WidgetStack.Stack.active = WidgetStack.Modules
-
-  table.insert(widgets, WidgetStack.Stack)
-
-  return widgets
-end
+-- init ./Modules/GuiWidgets
+local GuiWidgets = require 'Modules.GuiWidgets'
+GuiWidgets.Init(dt, LogHelper, Helper, Workflow, WidgetStack)
 
 ---------------------------------------------------------------
+
+local Env =
+{
+  InstallModuleEventRegistered = false,
+  InstallModuleDone = false,
+}
 
 -- register the module and create widget box in lighttable and darkroom
 local function InstallModuleRegisterLib()
@@ -418,7 +144,7 @@ local function InstallModuleRegisterLib()
       {
         orientation = 'vertical',
         reset_callback = SetAllDefaultModuleConfigurations,
-        table.unpack(GetWidgets()),
+        table.unpack(GuiWidgets.GetWidgets()),
       },
 
       nil, -- view_enter
